# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: |
    API untuk sistem marketplace dengan fitur:
    - Manajemen produk
    - Sistem reseller dengan komisi
    - Tracking click dan leads
    - Manajemen payout
    - Integrasi MinIO untuk upload file
  version: 1.0.0
  contact:
    name: API Support
    email: support@marketplace.com
servers:
  - url: http://localhost:8080/api
    description: Development Server
  - url: https://api.marketplace.com/api
    description: Production Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token yang didapat dari endpoint login
  schemas:
    Error:
      type: object
      properties:
        api_call_id:
          type: string
          example: "API_CALL_1234567890"
        message_action:
          type: string
          example: "PLEASE_TRY_AGAIN"
        message_data:
          type: string
          example: "Invalid request"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, RESELLER, CLIENT]
        referral_code:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/login/admin:
    post:
      tags:
        - Auth
      summary: Admin login
      description: Login untuk admin dan super admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@marketplace.com"
                password:
                  type: string
                  format: password
                  example: "admin123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_call_id:
                    type: string
                  message_action:
                    type: string
                    example: "LOGIN_SUCCESS"
                  message_data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT token
                      role:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login/reseller:
    post:
      tags:
        - Auth
      summary: Reseller login
      description: Login untuk reseller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - password
              properties:
                phone:
                  type: string
                  example: "081234567890"
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_call_id:
                    type: string
                  message_action:
                    type: string
                    example: "LOGIN_SUCCESS"
                  message_data:
                    type: object
                    properties:
                      token:
                        type: string
                      role:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'

  # ==================== ADMIN ENDPOINTS ====================
...